// packages/ldo-types/types/activitypub.typings.ts
export interface Object {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: string;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
}

/**
 * Link Type
 */
export interface Link {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: {
    "@id": "Link";
  };
  name?: string;
  preview?: LdSet<string>;
  height?: number;
  href?: string;
  hreflang?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  rel?: LdSet<string>;
  width?: number;
}

/**
 * Activity Type
 */
export interface Activity {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string>;
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  actor?: LdSet<string>;
  instrument?: LdSet<string>;
  object?: LdSet<string>;
  origin?: LdSet<string>;
  result?: LdSet<string>;
  target?: LdSet<string>;
}

/**
 * IntransitiveActivity Type
 */
export interface IntransitiveActivity {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string>;
  "@id"?: LdSet<string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  actor?: LdSet<string>;
  instrument?: LdSet<string>;
  object?: LdSet<string>;
  origin?: LdSet<string>;
  result?: LdSet<string>;
  target?: LdSet<string>;
}

/**
 * Collection Type
 */
export interface Collection {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string>;
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  totalItems?: number;
  current?: string;
  first?: string;
  items?: LdSet<string>;
  last?: string;
}

/**
 * OrderedCollection Type
 */
export interface OrderedCollection {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string>;
  "@id"?: LdSet<string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  totalItems?: number;
  current?: string;
  first?: string;
  items?: LdSet<string>;
  last?: string;
  orderedItems?: LdSet<string>;
}

/**
 * CollectionPage Type
 */
export interface CollectionPage {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string>;
  "@id"?: LdSet<string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  totalItems?: number;
  current?: string;
  first?: string;
  items?: LdSet<string>;
  last?: string;
  next?: string;
  partOf?: string;
  prev?: string;
}

/**
 * OrderedCollectionPage Type
 */
export interface OrderedCollectionPage {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string | string>;
  "@id"?: LdSet<string | string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  totalItems?: number;
  current?: string;
  first?: string;
  items?: LdSet<string>;
  last?: string;
  orderedItems?: LdSet<string>;
  startIndex?: number;
}

/**
 * AcceptShape Type
 */
export interface AcceptShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string>;
  "@id"?: LdSet<string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  actor?: LdSet<string>;
  instrument?: LdSet<string>;
  object?: LdSet<string>;
  origin?: LdSet<string>;
  result?: LdSet<string>;
  target?: LdSet<string>;
}

/**
 * AddShape Type
 */
export interface AddShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string>;
  "@id"?: LdSet<string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  actor?: LdSet<string>;
  instrument?: LdSet<string>;
  object?: LdSet<string>;
  origin?: LdSet<string>;
  result?: LdSet<string>;
  target?: LdSet<string>;
}

/**
 * CreateShape Type
 */
export interface CreateShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string>;
  "@id"?: LdSet<string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  actor?: LdSet<string>;
  instrument?: LdSet<string>;
  object?: LdSet<string>;
  origin?: LdSet<string>;
  result?: LdSet<string>;
  target?: LdSet<string>;
}

/**
 * DeleteShape Type
 */
export interface DeleteShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<string | string | string>;
  "@id"?: LdSet<string | string>;
  "@context"?: LdSet<LdoJsonldContext | LdoJsonldContext>;
  attachment?: LdSet<string>;
  attributedTo?: LdSet<string>;
  audience?: LdSet<string>;
  bcc?: LdSet<string>;
  bto?: LdSet<string>;
  cc?: LdSet<string>;
  content?: string;
  context?: LdSet<string>;
  generator?: LdSet<string>;
  icon?: LdSet<string>;
  image?: LdSet<string>;
  inReplyTo?: LdSet<string>;
  location?: LdSet<string>;
  name?: string;
  preview?: LdSet<string>;
  replies?: string;
  summary?: string;
  tag?: LdSet<string>;
  to?: LdSet<string>;
  url?: LdSet<string>;
  duration?: string;
  endTime?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  published?: string;
  startTime?: string;
  updated?: string;
  actor?: LdSet<string>;
  instrument?: LdSet<string>;
  object?: LdSet<string>;
  origin?: LdSet<string>;
  result?: LdSet<string>;
  target?: LdSet<string>;
}
// ...
export interface AudioShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: string;
}

/**
 * ImageShape Type
 */
export interface ImageShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: string;
}

/**
 * VideoShape Type
 */
export interface VideoShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: string;
}
// ...
export interface MentionShape {
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  type: LdSet<
    | string
    | {
        "@id": "Link";
      }
  >;
  "@id"?: string;
  "@context"?: LdoJsonldContext;
  name?: string;
  preview?: LdSet<string>;
  height?: number;
  href?: string;
  hreflang?: string;
  id?: {
    "@id": string;
  };
  mediaType?: string;
  rel?: LdSet<string>;
  width?: number;
}